--- how-to/register-with-browser/client/src/bootstrapper.ts
@@ -1,13 +1,0 @@
-import { createLaunchBarWindow } from "./launchbar-window";
-import { getSettings } from "./settings";
-
-export async function init() {
-	// you can kick off your bootstrapping process here where you may decide to prompt for authentication,
-	// gather reference data etc before starting workspace and interacting with it.
-	console.log("Initialising the bootstrapper");
-	const {
-		bootstrap: { launchBarWindowSettings }
-	} = await getSettings();
-
-	await createLaunchBarWindow(launchBarWindowSettings);
-}
