--- how-to/register-with-browser/client/src/launchbar-window.ts
@@ -1,5 +1,0 @@
-import type { LaunchBarWindowSettings } from "./shapes";
-
-export async function createLaunchBarWindow(options: LaunchBarWindowSettings) {
-	return fin.Window.create(options);
-}
