+++ how-to/register-with-browser/client/tsconfig.json
@@ -6,6 +6,7 @@
 		"rootDir": "./src",
 		"outDir": "build",
 		"skipLibCheck": true,
+		"strict": true,
 		"esModuleInterop": true,
 		"forceConsistentCasingInFileNames": true,
 		"importsNotUsedAsValues": "error",
