+++ how-to/register-with-browser/public/html/right-panel.html
@@ -8,7 +8,7 @@
 		<link rel="stylesheet" href="/common/style/app.css" />
 	</head>
 
-	<body class="col fill gap20 theme-light pad10 center">
+	<body class="col fill gap20 pad10 center">
 		<p>Right Panel</p>
 	</body>
 </html>
