+++ how-to/register-with-browser/public/manifest.fin.json
@@ -10,11 +10,7 @@
 		"icon": "http://localhost:8080/favicon.ico",
 		"autoShow": false,
 		"providerUrl": "http://localhost:8080/platform/provider.html",
-		"preventQuitOnLastWindowClosed": true,
-		"defaultWindowOptions": {}
-	},
-	"snapshot": {
-		"windows": []
+		"preventQuitOnLastWindowClosed": true
 	},
 	"shortcut": {
 		"company": "OpenFin",
@@ -30,72 +26,21 @@
 		"forwardErrorReports": true
 	},
 	"customSettings": {
-		"bootstrap": {
-			"launchBarWindowSettings": {
-				"url": "http://localhost:8080/html/launch-bar.html",
-				"name": "launch-bar",
-				"frame": false,
-				"icon": "https://www.openfin.co/favicon.ico",
-				"defaultHeight": 470,
-				"defaultWidth": 340,
-				"defaultCentered": true,
-				"saveWindowState": false,
-				"resizable": false,
-				"alwaysOnTop": true,
-				"cornerRounding": {
-					"height": 10,
-					"width": 10
-				}
+		"launchBarWindowSettings": {
+			"url": "http://localhost:8080/html/launch-bar.html",
+			"name": "launch-bar",
+			"frame": false,
+			"icon": "http://localhost:8080/favicon.ico",
+			"defaultHeight": 470,
+			"defaultWidth": 340,
+			"defaultCentered": true,
+			"saveWindowState": false,
+			"resizable": false,
+			"alwaysOnTop": true,
+			"cornerRounding": {
+				"height": 10,
+				"width": 10
 			}
-		},
-		"browserProvider": {
-			"windowOptions": {
-				"title": "Browser Starter",
-				"icon": "https://www.openfin.co/favicon.ico",
-				"newTabUrl": "http://localhost:8080/common/views/platform/new-tab/new-tab.html",
-				"newPageUrl": "http://localhost:8080/common/views/platform/new-tab/new-tab.html"
-			}
-		},
-		"themeProvider": {
-			"themes": [
-				{
-					"label": "Starter Theme",
-					"logoUrl": "https://www.openfin.co/favicon.ico",
-					"palette": {
-						"brandPrimary": "#1C355E",
-						"brandSecondary": "#0075BE",
-						"backgroundPrimary": "#FFFFFF",
-						"functional1": null,
-						"functional2": null,
-						"functional3": null,
-						"functional4": null,
-						"functional5": null,
-						"functional6": null,
-						"functional7": null,
-						"functional8": null,
-						"functional9": null,
-						"functional10": null,
-						"statusSuccess": null,
-						"statusWarning": null,
-						"statusCritical": null,
-						"statusActive": null,
-						"inputBackground": null,
-						"inputColor": null,
-						"inputPlaceholder": null,
-						"inputDisabled": null,
-						"inputFocused": null,
-						"textDefault": null,
-						"textHelp": null,
-						"textInactive": null,
-						"background1": "#1C355E",
-						"background2": "#0075BE",
-						"background3": "#1C355E",
-						"background4": "#1C355E",
-						"background5": "#0075BE",
-						"background6": "#0075BE"
-					}
-				}
-			]
 		}
 	}
 }
